docker_scripts/train_readernet_v5/requirements.txt
supabase
python-dotenv
requests
pillow
pillow-avif-plugin
onnx
onnxruntime
huggingface_hub
tqdm
scikit-learn
